move drawing styles to the drawing module
