+++ src/components/analyzers/basic-analyzer/front-view/home-button.tsx
@@ -1,5 +1,5 @@
 import classNames from "classnames";
-import drawingStyles from "../../drawing.module.css";
+import drawingStyles from "../../drawing/drawing.module.css";
 import { ButtonProps } from "../button-props";
 import { PositionableProps } from "../../drawing";
