+++ src/components/analyzers/drawing/drawing.module.css
