+++ src/components/analyzers/drawing/index.ts
@@ -1,1 +1,2 @@
 export { PositionableProps } from "./positionable-props";
+// export { default as drawingStyles } from "./joystick.module.css";
